Adding reliability analysis code files
All files required to run the reliability.ipynb file. These files were used to create the coding analysis of the reliability section of the report. This section is assigned to Isabella Tranquilino (s4005248).

A placeholder for youtube v3 API key is inside the .ipynb file, an API key is required to run it.